split NodeHelpers to remove circular dependencies
